requirements.txt
newspaper3k==0.2.8
PyYAML==6.0.1
Requests==2.31.0
spacy==3.7.2
